more work on reducing warnings.

+++ freex3d/src/lib/input/SensInterps.c
@@ -45,7 +45,6 @@ Interps are the "EventsProcessed" fields of interpolators.
 #include "opengl/LoadTextures.h"        /* for finding a texture url in a multi url */
 
 
-
 #include <resources.h>
 
 #include "../main/headers.h"
@@ -57,6 +56,7 @@ Interps are the "EventsProcessed" fields of interpolators.
 #include "../scenegraph/quaternion.h"
 #include "../scenegraph/sounds.h"
 #include "../vrml_parser/CRoutes.h"
+#include "../opengl/OpenGL_Utils.h"
 
 #include "SensInterps.h"
 
@@ -1645,7 +1645,7 @@ void do_SphereSensor ( void *ptr, int ev, int but1, int over) {
 			printf ("warning, newRad %lf == 0, can not compute\n",newRad);
 			return;
 		}
-		RADIUS = newRad;
+		RADIUS = (float) newRad;
 
 		/* save the current norm here */
 		NORM_CUR_X = CUR_X / RADIUS;
